update optimization.py, move reweight_ct
